+++ NEWS.md
@@ -1,3 +1,8 @@
+# Keng development
+
+* test_r() added Fisher's z.
+* compare_lm() added *f_squared* and post-hoc power.
+
 # Keng 2024.11.17
 
 * Added a `NEWS.md` file to track changes to the package.